Add files via upload
Here is my Python project. It is a code for an ATM and allows the user to withdraw, deposit or check their balance. The balance is £28.07 and the PIN is 1234. The code ends if the wrong PIN is typed in. When the user is withdrawing or depositing money, the new balance is shown.